test(step3): 랜덤 Generator 테스트로 클래스 분리
